Replaced switch_hidden with refactor file
